remove ui/component/widget
